Value extends Foldable
