+++ src/test/java/javaslang/control/OptionTest.java
@@ -6,6 +6,8 @@
 package javaslang.control;
 
 import javaslang.Serializables;
+import javaslang.Value;
+import javaslang.collection.AbstractValueTest;
 import org.junit.Test;
 
 import java.util.*;
